adk08/requirements.txt
google-cloud-bigquery
google-auth
python-dotenv
requests
google-adk
